% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumBoth.R
\name{SumBoth}
\alias{SumBoth}
\title{SumBoth}
\usage{
SumBoth(num1, num2)
}
\arguments{
\item{num1}{First number}

\item{num2}{Second number}
}
\value{
The sum of \code{num1} and \code{num2}

This function takes two real numbers and returns their summation
}
\description{
A function that takes two real numbers and returns their summation.
}
\examples{

#' # Sum two known numbers
x <- 1
y <- 2
# x + y = 1 + 2 = 3
z <- SumBoth(x, y)
z

# What happens if input is not from numeric class
x <- "a"
y <- 2
SumBoth(x, y)

}
\author{
Jaime Mosquera Gutiérrez, \email{jmosquerag@unal.edu.co}
}
